improve: PR 7 개선사항 적용
- as any를 적절한 타입으로 교체
- Alert, Dialog 컴포넌트를 개별 파일로 분리하여 폴더 구조화
- AppGuard가 전체 앱을 감싸도록 위치 수정
- renderer 파일들의 import문에서 확장자 제거
