+++ apps/desktop/src/renderer/common/Alert/AlertDescription.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import cn from '../shared/cn'
+
+const AlertDescription = React.forwardRef<
+  HTMLParagraphElement,
+  React.HTMLAttributes<HTMLParagraphElement>
+>(({ className, ...props }, ref) => (
+  <div
+    ref={ref}
+    className={cn('text-sm [&_p]:leading-relaxed', className)}
+    {...props}
+  />
+))
+AlertDescription.displayName = 'AlertDescription'
+
+export { AlertDescription }
