+++ apps/desktop/src/renderer/common/Alert/AlertTitle.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import cn from '../shared/cn'
+
+const AlertTitle = React.forwardRef<
+  HTMLParagraphElement,
+  React.HTMLAttributes<HTMLHeadingElement>
+>(({ className, ...props }, ref) => (
+  <h5
+    ref={ref}
+    className={cn('mb-1 font-medium leading-none tracking-tight', className)}
+    {...props}
+  />
+))
+AlertTitle.displayName = 'AlertTitle'
+
+export { AlertTitle }
