+++ apps/desktop/src/renderer/common/Alert/index.tsx
@@ -1,0 +1,3 @@
+export { Alert } from './Alert'
+export { AlertTitle } from './AlertTitle'
+export { AlertDescription } from './AlertDescription'
