+++ apps/desktop/src/renderer/common/Dialog/DialogFooter.tsx
@@ -1,0 +1,18 @@
+import * as React from 'react'
+import cn from '../shared/cn'
+
+const DialogFooter = ({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) => (
+  <div
+    className={cn(
+      'flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2',
+      className
+    )}
+    {...props}
+  />
+)
+DialogFooter.displayName = 'DialogFooter'
+
+export { DialogFooter }
