+++ apps/desktop/src/renderer/common/Dialog/DialogRoot.tsx
@@ -1,0 +1,9 @@
+import * as DialogPrimitive from '@radix-ui/react-dialog'
+
+// Re-export Radix primitives
+const Dialog = DialogPrimitive.Root
+const DialogTrigger = DialogPrimitive.Trigger
+const DialogPortal = DialogPrimitive.Portal
+const DialogClose = DialogPrimitive.Close
+
+export { Dialog, DialogTrigger, DialogPortal, DialogClose }
