+++ apps/desktop/src/renderer/pages/AppGuard/StorageLocationDialog/ErrorMessage.tsx
@@ -1,4 +1,4 @@
-import { Alert, AlertDescription } from '@/renderer/common/Alert.js'
+import { Alert, AlertDescription } from '@/renderer/common/Alert'
 
 interface ErrorMessageProps {
   error: unknown
@@ -6,9 +6,9 @@ interface ErrorMessageProps {
 
 export default function ErrorMessage({ error }: ErrorMessageProps) {
   if (!error) return null
-  
+
   const message = error instanceof Error ? error.message : '오류가 발생했습니다'
-  
+
   return (
     <Alert variant="destructive">
       <AlertDescription>{message}</AlertDescription>
