+++ apps/desktop/src/renderer/pages/AppGuard/StorageLocationDialog/index.tsx
@@ -7,11 +7,11 @@ import {
   DialogFooter,
   DialogHeader,
   DialogTitle,
-} from '@/renderer/common/Dialog.js'
-import { Button } from '@/renderer/common/Button.js'
-import InitializingView from './InitializingView.js'
-import WorkspacePathSelector from './WorkspacePathSelector.js'
-import ErrorMessage from './ErrorMessage.js'
+} from '@/renderer/common/Dialog'
+import { Button } from '@/renderer/common/Button'
+import InitializingView from './InitializingView'
+import WorkspacePathSelector from './WorkspacePathSelector'
+import ErrorMessage from './ErrorMessage'
 
 export default function StorageLocationDialog() {
   const [selectedPath, setSelectedPath] = useState('')
@@ -28,32 +28,23 @@ export default function StorageLocationDialog() {
 
   return (
     <Dialog open={true}>
-      <DialogContent 
-        className="sm:max-w-[500px]" 
-        onPointerDownOutside={(e) => e.preventDefault()}
-      >
+      <DialogContent className="sm:max-w-[500px]" onPointerDownOutside={(e) => e.preventDefault()}>
         <DialogHeader>
           <DialogTitle>Voyl 워크스페이스 설정</DialogTitle>
           <DialogDescription>
-            데이터를 저장할 폴더를 선택해주세요. 
-            Google Drive나 Dropbox 폴더를 선택하면 여러 기기에서 동기화할 수 있습니다.
+            데이터를 저장할 폴더를 선택해주세요. Google Drive나 Dropbox 폴더를 선택하면 여러
+            기기에서 동기화할 수 있습니다.
           </DialogDescription>
         </DialogHeader>
 
         <div className="grid gap-4 py-4">
-          <WorkspacePathSelector 
-            selectedPath={selectedPath}
-            onPathSelect={setSelectedPath}
-          />
-          
+          <WorkspacePathSelector selectedPath={selectedPath} onPathSelect={setSelectedPath} />
+
           <ErrorMessage error={initialize.error} />
         </div>
 
         <DialogFooter>
-          <Button
-            onClick={handleInitialize}
-            disabled={!selectedPath}
-          >
+          <Button onClick={handleInitialize} disabled={!selectedPath}>
             워크스페이스 생성
           </Button>
         </DialogFooter>
